+++ Targets/D3N/Sources/Feature/TodayNews/Main/Cell/TodayNewsListItemCellStore.swift
@@ -1,8 +1,8 @@
 //
-//  TodayListItemCellStore.swift
+//  TodayNewsListItemCellStore.swift
 //  D3N
 //
-//  Created by 송영모 on 11/6/23.
+//  Created by 송영모 on 11/30/23.
 //  Copyright © 2023 sju. All rights reserved.
 //
 
@@ -10,17 +10,17 @@ import Foundation
 
 import ComposableArchitecture
 
-public struct TodayListItemCellStore: Reducer {
+public struct TodayNewsListItemCellStore: Reducer {
     public struct State: Equatable, Identifiable {
         public var id: UUID
-        public var newsEntity: NewsEntity
+        public var todayNews: TodayNewsEntity
         
         public init(
             id: UUID = .init(),
-            newsEntity: NewsEntity
+            todayNews: TodayNewsEntity
         ) {
             self.id = id
-            self.newsEntity = newsEntity
+            self.todayNews = todayNews
         }
     }
